+++ src/index.js
@@ -71,6 +71,7 @@ class WithObservablesComponent<AddedValues: any, PropsInput: {}> extends Compone
   {
     isFetching: boolean,
     values: $FlowFixMe<AddedValues>,
+    error: ?Error,
     triggeredFromProps: any[],
   },
 > {
@@ -101,6 +102,7 @@ class WithObservablesComponent<AddedValues: any, PropsInput: {}> extends Compone
     this.state = {
       isFetching: true,
       values: {},
+      error: null,
       triggeredFromProps: getTriggeringProps(props, triggerProps),
     }
 
@@ -117,7 +119,7 @@ class WithObservablesComponent<AddedValues: any, PropsInput: {}> extends Compone
 
     scheduleForCleanup(() => {
       if (!this._prefetchTimeoutCanceled) {
-        console.warn(`withObservables - unsubscribing from source. Leaky component!`)
+        console.warn(`[withObservables] Unsubscribing from source. Leaky component!`)
         this.unsubscribe()
       }
     })
@@ -131,7 +133,7 @@ class WithObservablesComponent<AddedValues: any, PropsInput: {}> extends Compone
 
     if (!this._subscription) {
       console.warn(
-        `withObservables - component mounted but no subscription present. Slow component (timed out) or something weird happened! Re-subscribing`,
+        `[withObservables] Component mounted but no subscription present. Slow component (timed out) or a bug! Re-subscribing...`,
       )
 
       const newTriggeringProps = getTriggeringProps(this.props, this.triggerProps)
@@ -163,11 +165,7 @@ class WithObservablesComponent<AddedValues: any, PropsInput: {}> extends Compone
     this.unsubscribe()
     this._subscription = this.getNewProps(props).subscribe(
       values => this.withObservablesOnChange(values),
-      error => {
-        // we need to explicitly log errors from the new observables, or they will get lost
-        // TODO: It can be difficult to trace back the component in which this error originates. We should maybe propagate this as an error of the component? Or at least show in the error a reference to the component, or the original `getProps` function?
-        console.error(`Error in Rx composition in withObservables()`, error)
-      },
+      error => this.withObservablesOnError(error),
     )
   }
 
@@ -185,6 +183,23 @@ class WithObservablesComponent<AddedValues: any, PropsInput: {}> extends Compone
       this.state.values = values
       this.state.isFetching = false
     }
+  }
+
+  // DO NOT rename (we want on call stack as debugging help)
+  withObservablesOnError(error: Error): void {
+    // console.error(`[withObservables] Error in Rx composition`, error)
+    if (this._exitedConstructor) {
+      this.setState({
+        error,
+        isFetching: false,
+      })
+    } else {
+      this.state.error = error
+      this.state.isFetching = false
+    }
+    // TODO: Unsubscribe proactively to lessen GC pressure
+    // since we know for sure we won't be able to use the Observable again
+    // this.unsubscribe()
   }
 
   unsubscribe(): void {
@@ -207,10 +222,17 @@ class WithObservablesComponent<AddedValues: any, PropsInput: {}> extends Compone
   }
 
   render(): * {
-    const { isFetching, values } = this.state
-    return isFetching ?
-      null :
-      createElement(this.BaseComponent, Object.assign({}, this.props, values))
+    const { isFetching, values, error } = this.state
+
+    if (isFetching) {
+      return null
+    } else if (error) {
+      // rethrow error found in Rx composition as to unify withObservables errors with other React errors
+      // the responsibility for handling errors is on the user (by using an Error Boundary)
+      throw error
+    } else {
+      return createElement(this.BaseComponent, Object.assign({}, this.props, values))
+    }
   }
 }
 
@@ -229,6 +251,8 @@ class WithObservablesComponent<AddedValues: any, PropsInput: {}> extends Compone
 //
 // If you only want to subscribe to Observables once (the Observables don't depend on outer props),
 // pass `null` to `triggerProps`.
+//
+// Errors are re-thrown in render(). Use React Error Boundary to catch them.
 //
 // Example use:
 //   withObservablesSynchronized(['task'], ({ task }) => ({